requirements.txt
git+https://github.com/openai/whisper.git
flask
whisper
torch
ffmpeg-python
deep-translator
python-dotenv